Clean up. Upgrade Visual Studio solution to VS2017. Fix compiler warnings.
TODO: Compiles but doesn't run.

+++ software-renderer/crender-mt/api-command.h
@@ -120,7 +120,7 @@ namespace Ceng
 	protected:
 
 		Ceng::FLOAT32 depth;
-		Ceng::FLOAT32 stencil;
+		Ceng::UINT32 stencil;
 
 	public:
 		Cmd_ClearDepthStencil();
